Conn improvements
- Reduce Station power to LOW
- Remove all Thr queue code, just resend last command every 100ms

+++ Throttle/LocoState.h
@@ -25,8 +25,8 @@ extern struct Loco loco;
 struct Controls {
   uint16_t throttle;
   int16_t direction;
-  int inbound;
-  int outbound;
+  float inbound;
+  float outbound;
 };
 extern struct Controls controls;
